+++ client/src/pages/index.ts
@@ -2,4 +2,4 @@ export * from "./landing";
 export * from "./home";
 export * from "./event";
 export * from "./apply-form";
-export * from "./notFound";
+export * from "./not-found";
